feat: add exception handling and restrictions for some variables for Items and Categories

+++ src/main/java/com/mita/service/CategoryService.java
@@ -50,6 +50,10 @@ public class CategoryService {
     }
 
     public void deleteCategoryById(Long id) {
+        if (!categoryRepository.existsById(id)) {
+            throw new IllegalArgumentException("Category with id: " + id + " not found");
+        }
+
         categoryRepository.deleteById(id);
     }
